template placing overlapping with world border for more interesting stuff along edges

+++ inc/Common.h
@@ -13,13 +13,13 @@ enum class MovementDirection
 
 struct Position
 {
-	unsigned int x;
-	unsigned int y;
+	int x;
+	int y;
 
-	Position(unsigned int x, unsigned int y) : x(x), y(y) {}
+	Position(int x, int y) : x(x), y(y) {}
 	Position() : x(-1), y(-1) {} // purposeful overflow, max val to indicate invalidity
 
-	bool isValid(unsigned int maxX, unsigned int maxY) const
+	bool isValid(int maxX, int maxY) const
 	{
 		return x < maxX && y < maxY;
 	}
@@ -53,7 +53,7 @@ struct Position
 	{
 		return {x + other.x, y + other.y};
 	}
-	Position operator+(unsigned int scalar) const
+	Position operator+(int scalar) const
 	{
 		return {x + scalar, y + scalar};
 	}
@@ -61,14 +61,14 @@ struct Position
 	{
 		return {x - other.x, y - other.y};
 	}
-	Position operator*(unsigned int scalar) const
+	Position operator*(int scalar) const
 	{
 		return {x * scalar, y * scalar};
 	}
 
 	double distance(const Position& other) const
 	{
-		return std::abs(static_cast<int>(x) - static_cast<int>(other.x)) + std::abs(static_cast<int>(y) - static_cast<int>(other.y));
+		return std::abs(x - other.x) + std::abs(y - other.y);
 	}
 };
